Move command_line_assistant.initialize
Rename command_line_assistant.initialize to command_line_assistant.client
