+++ data/release/man/c.1
@@ -27,7 +27,7 @@ level margin: \\n[rst2man-indent\\n[rst2man-indent-level]]
 .\" new: \\n[rst2man-indent\\n[rst2man-indent-level]]
 .in \\n[rst2man-indent\\n[rst2man-indent-level]]u
 ..
-.TH "C" "1" "Apr 03, 2025" "0.3.1" "Command Line Assistant"
+.TH "C" "1" "Apr 08, 2025" "0.3.1" "Command Line Assistant"
 .SH NAME
 c \- Command Line Assistant Client
 .SH SYNOPSIS
@@ -73,7 +73,7 @@ c chat [\-h] [\-a [ATTACHMENT]] [\-i] [\-w [WITH_OUTPUT]] [\-r] [\-l] [\-d [DELE
 .IP \(bu 2
 \fI\%query_string\fP \- The question that will be sent to the LLM (default: \fB\fP)
 .IP \(bu 2
-\fI\%\-a\fP \fBATTACHMENT\fP, \fI\%\-\-attachment\fP \fBATTACHMENT\fP \- File attachment to be read and sent alongside the query (default: \fB\fP)
+\fI\%\-a\fP \fBATTACHMENT\fP, \fI\%\-\-attachment\fP \fBATTACHMENT\fP \- File attachment to be read and sent alongside the query (default: \fBNone\fP)
 .IP \(bu 2
 \fI\%\-i\fP, \fI\%\-\-interactive\fP \- Start an interactive chat session
 .IP \(bu 2
