+++ docs/source/client.rst
@@ -1,0 +1,8 @@
+Client
+======
+
+.. automodule:: command_line_assistant.client
+   :members:
+   :undoc-members:
+   :private-members:
+   :no-index:
