--- docs/source/initialize.rst
@@ -1,8 +1,0 @@
-Initialize
-==========
-
-.. automodule:: command_line_assistant.initialize
-   :members:
-   :undoc-members:
-   :private-members:
-   :no-index:
