+++ pyproject.toml
@@ -21,7 +21,7 @@ readme = "README.md"
 license = { file = "LICENSE" }
 classifiers = ["Programming Language :: Python :: 3"]
 urls = { Repository = "https://github.com/rhel-lightspeed/command-line-assistant.git", Issues = "https://github.com/rhel-lightspeed/command-line-assistant/issues" }
-scripts = { c = "command_line_assistant.initialize:initialize", clad = "command_line_assistant.daemon.clad:daemonize" }
+scripts = { c = "command_line_assistant.client:initialize", clad = "command_line_assistant.daemon.clad:daemonize" }
 
 [build-system]
 # poetry build is not available in rhel baseos repositories
